fix change data type

+++ features/feedback/handler/request.go
@@ -1,9 +1,9 @@
 package handler
 
 type NewFeedbackRequest struct {
-	RoomID    uint	`json:"room_id"`
+	RoomID    uint `json:"room_id"`
 	UserID    uint
 	CreatedAt string
-	Rating    int    `json:"rating"`
-	Comment   string `json:"comment"`
+	Rating    float64 `json:"rating"`
+	Comment   string  `json:"comment"`
 }
