+++ features/feedback/handler/response.go
@@ -1,11 +1,11 @@
 package handler
 
 type FeedbackResponse struct {
-	ID        uint   `json:"id"`
-	UserName  string `json:"user_name"`
-	Rating    int    `json:"rating"`
-	Comment   string `json:"comment"`
-	CreatedAt string `json:"created_at"`
+	ID        uint    `json:"id"`
+	UserName  string  `json:"user_name"`
+	Rating    float64 `json:"rating"`
+	Comment   string  `json:"comment"`
+	CreatedAt string  `json:"created_at"`
 }
 
 type ListFeedbackResponse []FeedbackResponse
